+++ clients/extension/src/inpage/index.ts
@@ -1,41 +1,6 @@
-import {
-  CosmosMsgType,
-  MessageKey,
-  RequestMethod,
-} from '@clients/extension/src/utils/constants'
-import { getCosmosChainFromAddress } from '@clients/extension/src/utils/cosmos/getCosmosChainFromAddress'
-import {
-  Messaging,
-  SendTransactionResponse,
-  TransactionDetails,
-  TransactionType,
-  VaultExport,
-} from '@clients/extension/src/utils/interfaces'
-import { CosmosChain } from '@core/chain/Chain'
-import { getCosmosAccountInfo } from '@core/chain/chains/cosmos/account/getCosmosAccountInfo'
-import { chainFeeCoin } from '@core/chain/coin/chainFeeCoin'
-import { getChainByChainId } from '@core/chain/coin/ChainId'
-import {
-  CosmJSOfflineSigner,
-  CosmJSOfflineSignerOnlyAmino,
-  Keplr,
-} from '@keplr-wallet/provider'
-import {
-  AminoSignResponse,
-  BroadcastMode,
-  DirectSignResponse,
-  KeplrMode,
-  KeplrSignOptions,
-  Key,
-  OfflineAminoSigner,
-  OfflineDirectSigner,
-  StdSignDoc,
-  StdTx,
-} from '@keplr-wallet/types'
-import base58 from 'bs58'
-import { TxBody } from 'cosmjs-types/cosmos/tx/v1beta1/tx'
-import { MsgTransfer } from 'cosmjs-types/ibc/applications/transfer/v1/tx'
-import Long from 'long'
+import { MessageKey } from '@clients/extension/src/utils/constants'
+import { Messaging, VaultExport } from '@clients/extension/src/utils/interfaces'
+
 import { announceProvider, EIP1193Provider } from 'mipd'
 import { v4 as uuidv4 } from 'uuid'
 
@@ -50,303 +15,12 @@ import { Solana } from './providers/solana'
 import { THORChain } from './providers/thorchain'
 import { UTXO } from './providers/utxo'
 import { shouldInjectProvider } from './utils/injectHelpers'
+import { XDEFIKeplrProvider } from './providers/xdefiKeplr'
 
 export type Callback = (
   error: Error | null,
   result?: Messaging.Chain.Response
 ) => void
-
-class XDEFIMessageRequester {
-  constructor() {
-    this.sendMessage = this.sendMessage.bind(this)
-  }
-  // Expected for Ctrl
-  public async sendMessage(_message: any, _params: any): Promise<void> {
-    return new Promise(resolve => {
-      resolve()
-    })
-  }
-}
-
-class XDEFIKeplrProvider extends Keplr {
-  static instance: XDEFIKeplrProvider | null = null
-  isXDEFI: boolean
-  isVulticonnect: boolean
-
-  public static getInstance(): XDEFIKeplrProvider {
-    if (!XDEFIKeplrProvider.instance) {
-      XDEFIKeplrProvider.instance = new XDEFIKeplrProvider(
-        '0.0.1',
-        'extension',
-        new XDEFIMessageRequester()
-      )
-    }
-    return XDEFIKeplrProvider.instance
-  }
-
-  emitAccountsChanged(): void {
-    window.dispatchEvent(new Event('keplr_keystorechange'))
-  }
-
-  constructor(version: string, mode: KeplrMode, requester: any) {
-    super(version, mode, requester)
-    this.isXDEFI = true
-    this.isVulticonnect = true
-    window.ctrlKeplrProviders = {}
-    window.ctrlKeplrProviders['Ctrl Wallet'] = this
-  }
-  enable(_chainIds: string | string[]): Promise<void> {
-    return new Promise<void>((resolve, reject) => {
-      cosmosProvider
-        .request({
-          method: RequestMethod.VULTISIG.REQUEST_ACCOUNTS,
-          params: [],
-        })
-        .then(resolve)
-        .catch(reject)
-    })
-  }
-  getOfflineSigner(
-    chainId: string,
-    _signOptions?: KeplrSignOptions
-  ): OfflineAminoSigner & OfflineDirectSigner {
-    const cosmSigner = new CosmJSOfflineSigner(
-      chainId,
-      window.xfi.keplr,
-      _signOptions
-    )
-
-    cosmSigner.getAccounts = async () => {
-      return cosmosProvider
-        .request({ method: RequestMethod.VULTISIG.CHAIN_ID, params: [] })
-        .then(async currentChainID => {
-          if (currentChainID !== chainId) {
-            return await cosmosProvider
-              .request({
-                method: RequestMethod.VULTISIG.WALLET_SWITCH_CHAIN,
-                params: [{ chainId }],
-              })
-              .then(async () => {
-                return await cosmosProvider.request({
-                  method: RequestMethod.VULTISIG.REQUEST_ACCOUNTS,
-                  params: [],
-                })
-              })
-          } else {
-            return await cosmosProvider.request({
-              method: RequestMethod.VULTISIG.REQUEST_ACCOUNTS,
-              params: [],
-            })
-          }
-        })
-    }
-
-    return cosmSigner as OfflineAminoSigner & OfflineDirectSigner
-  }
-
-  getOfflineSignerOnlyAmino(
-    chainId: string,
-    signOptions?: KeplrSignOptions
-  ): OfflineAminoSigner {
-    const cosmSigner = new CosmJSOfflineSignerOnlyAmino(
-      chainId,
-      window.xfi.keplr,
-      signOptions
-    )
-
-    cosmSigner.getAccounts = async () => {
-      return cosmosProvider
-        .request({ method: RequestMethod.VULTISIG.CHAIN_ID, params: [] })
-        .then(async currentChainID => {
-          if (currentChainID !== chainId) {
-            return await cosmosProvider
-              .request({
-                method: RequestMethod.VULTISIG.WALLET_SWITCH_CHAIN,
-                params: [{ chainId }],
-              })
-              .then(async () => {
-                return await cosmosProvider.request({
-                  method: RequestMethod.VULTISIG.REQUEST_ACCOUNTS,
-                  params: [],
-                })
-              })
-          } else {
-            return await cosmosProvider.request({
-              method: RequestMethod.VULTISIG.REQUEST_ACCOUNTS,
-              params: [],
-            })
-          }
-        })
-    }
-
-    return cosmSigner as OfflineAminoSigner
-  }
-
-  sendTx(
-    _chainId: string,
-    _tx: StdTx | Uint8Array,
-    _mode: BroadcastMode
-  ): Promise<Uint8Array> {
-    return new Promise<Uint8Array>((resolve, reject) => {
-      cosmosProvider
-        .request({
-          method: RequestMethod.VULTISIG.SEND_TRANSACTION,
-          params: [{ ..._tx, txType: 'Keplr' }],
-        })
-        .then((result: SendTransactionResponse) => {
-          const decoded = base58.decode(result.raw)
-          if (decoded) resolve(decoded)
-          else reject()
-        })
-        .catch(reject)
-    })
-  }
-  async sendMessage() {}
-
-  signAmino(
-    _chainId: string,
-    _signer: string,
-    signDoc: StdSignDoc,
-    _signOptions?: KeplrSignOptions
-  ): Promise<AminoSignResponse> {
-    return new Promise<AminoSignResponse>(resolve => {
-      const txDetails = signDoc.msgs.map(msg => {
-        if (msg.type === CosmosMsgType.MSG_SEND) {
-          return { txType: 'Keplr', ...msg.value } as TransactionType.Keplr
-        }
-      })
-
-      cosmosProvider
-        .request({
-          method: RequestMethod.VULTISIG.SEND_TRANSACTION,
-          params: [{ ...txDetails[0]!, txType: 'Keplr' }],
-        })
-        .then((result: SendTransactionResponse) => {
-          resolve(result as any)
-        })
-    })
-  }
-  async signDirect(
-    chainId: string,
-    signer: string,
-    signDoc: {
-      bodyBytes: Uint8Array
-      authInfoBytes: Uint8Array
-      chainId: string
-      accountNumber: Long
-    },
-    _signOptions: KeplrSignOptions
-  ): Promise<DirectSignResponse> {
-    const txBody = TxBody.decode(signDoc.bodyBytes)
-
-    const [firstMessage] = txBody.messages
-    if (
-      !firstMessage ||
-      firstMessage.typeUrl !== '/ibc.applications.transfer.v1.MsgTransfer'
-    ) {
-      throw new Error('Unsupported message type')
-    }
-
-    const txChain = getChainByChainId(chainId)
-    if (!txChain) {
-      throw new Error(`Chain not supported: ${chainId}`)
-    }
-
-    const msg = MsgTransfer.decode(firstMessage.value)
-    const receiverChain = getCosmosChainFromAddress(msg.receiver)
-    if (!receiverChain) {
-      throw new Error(`Receiver chain not supported: ${msg.receiver}`)
-    }
-
-    const standardTx: TransactionDetails = {
-      asset: {
-        chain: txChain,
-        ticker: msg.token.denom,
-      },
-      amount: {
-        amount: msg.token.amount,
-        decimals: chainFeeCoin[txChain].decimals,
-      },
-      from: msg.sender,
-      to: msg.receiver,
-      data: `${receiverChain}:${msg.sourceChannel}:${msg.receiver}:${msg.memo}`,
-      ibcTransaction: {
-        ...msg,
-        timeoutHeight: {
-          revisionHeight: msg.timeoutHeight.revisionHeight.toString(),
-          revisionNumber: msg.timeoutHeight.revisionNumber.toString(),
-        },
-        timeoutTimestamp: msg.timeoutTimestamp.toString(),
-      },
-    }
-
-    const result: SendTransactionResponse = await cosmosProvider.request({
-      method: RequestMethod.VULTISIG.SEND_TRANSACTION,
-      params: [{ ...standardTx, txType: 'Vultisig' }],
-    })
-
-    const accountInfo = await getCosmosAccountInfo({
-      chain: txChain as CosmosChain,
-      address: signer,
-    })
-
-    if (!accountInfo || !accountInfo.pubkey) {
-      throw new Error('No account info or pubkey')
-    }
-
-    const decoded = base58.decode(result.raw)
-    if (!decoded) {
-      throw new Error('Invalid signature')
-    }
-
-    return {
-      signed: signDoc,
-      signature: {
-        pub_key: accountInfo.pubkey,
-        signature: decoded.toString(),
-      },
-    }
-  }
-  async experimentalSuggestChain(_chainInfo: any) {
-    return
-  }
-  async sendSimpleMessage(
-    _type: string,
-    _method: string,
-    _payload: any
-  ): Promise<any> {
-    return
-  }
-
-  async getKey(chainId: string): Promise<Key> {
-    return cosmosProvider
-      .request({ method: RequestMethod.VULTISIG.CHAIN_ID, params: [] })
-      .then(async currentChainID => {
-        if (currentChainID !== chainId) {
-          return await cosmosProvider
-            .request({
-              method: RequestMethod.VULTISIG.WALLET_SWITCH_CHAIN,
-              params: [{ chainId }],
-            })
-            .then(async () => {
-              return (
-                await cosmosProvider.request({
-                  method: RequestMethod.VULTISIG.REQUEST_ACCOUNTS,
-                  params: [],
-                })
-              )[0]
-            })
-        } else {
-          return (
-            await cosmosProvider.request({
-              method: RequestMethod.VULTISIG.REQUEST_ACCOUNTS,
-              params: [],
-            })
-          )[0]
-        }
-      })
-  }
-}
 
 const bitcoinProvider = new UTXO(
   MessageKey.BITCOIN_REQUEST,
@@ -356,7 +30,7 @@ const bitcoinCashProvider = new UTXO(
   MessageKey.BITCOIN_CASH_REQUEST,
   'Bitcoincash_bitcoincash'
 )
-const cosmosProvider = new Cosmos()
+export const cosmosProvider = new Cosmos()
 const dashProvider = new Dash()
 const dogecoinProvider = new UTXO(
   MessageKey.DOGECOIN_REQUEST,
